% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDDchangepointWBS.R
\name{argmaxDelta}
\alias{argmaxDelta}
\title{This function calculates the argmax (change point candidate) and colsums of the difference distance matrix Delta for given data and dissimilarity measure}
\usage{
argmaxDelta(data, FUN = dist1)
}
\arguments{
\item{data}{a matrix of size nxp (i.e., n observations, each p-dimensional)}

\item{FUN}{a dissimilarity measure, default is the new dissimilarity measure d(Xi,Xj)}
}
\value{
the change point candidate (argmax) and colsums of the difference distance matrix Delta
}
\description{
This function calculates the argmax (change point candidate) and colsums of the difference distance matrix Delta for given data and dissimilarity measure
}
