% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDDchangepoint.R
\name{CI_changepoint}
\alias{CI_changepoint}
\title{This function calculates confidence intervals (CI) for multiple change points in high dimensional data}
\usage{
CI_changepoint(
  data,
  FUN = dist1,
  cp_estimate,
  npermutCI = 100,
  siglevel = 0.05
)
}
\arguments{
\item{data}{a matrix of size nxp (i.e., n observations, each p-dimensional)}

\item{FUN}{a dissimilarity measure, default is the new dissimilarity measure d(Xi,Xj)}

\item{cp_estimate}{a vector of change point locations}

\item{npermutCI}{number of permutations for permutation-based confidence intervals, default is 100}

\item{siglevel}{the significance level, default is 0.05}
}
\value{
the lower limit and upper limit of confidence intervals for the change point locations
}
\description{
This function calculates confidence intervals (CI) for multiple change points in high dimensional data
}
