% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDDchangepoint.R
\name{dist1}
\alias{dist1}
\title{This function uses the new dissimilarity measure d(Xi,Xj) to calculate the distance matrix of given data}
\usage{
dist1(data)
}
\arguments{
\item{data}{a matrix of size nxp (i.e., n observations, each p-dimensional)}
}
\value{
the distance matrix of data
}
\description{
This function uses the new dissimilarity measure d(Xi,Xj) to calculate the distance matrix of given data
}
