% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDDchangepointWBS.R
\name{multiple_changepoint_detection_wbs}
\alias{multiple_changepoint_detection_wbs}
\title{This is the function that runs the proposed algorithm with wild binary segmentation for multiple change point detection in high dimensional data}
\usage{
multiple_changepoint_detection_wbs(
  data,
  FUN = dist1,
  minsegmlength = 10,
  M = 50,
  test = "asymptotic",
  npermut = 200,
  siglevel = 0.05
)
}
\arguments{
\item{data}{a matrix of size nxp (i.e., n observations, each p-dimensional)}

\item{FUN}{a dissimilarity measure, default is the new dissimilarity measure d(Xi,Xj)}

\item{minsegmlength}{minimum segment length for wild binary segmentation, default is 10}

\item{M}{number of random draws for wild binary segmentation, default is 50}

\item{test}{either asymptotic or permutation test, default is the asymptotic test}

\item{npermut}{number of permutations for permutation test, default is 200}

\item{siglevel}{the significance level, default is 0.05}
}
\value{
the list of detected significant change points and the corresponding p-values
}
\description{
This is the function that runs the proposed algorithm with wild binary segmentation for multiple change point detection in high dimensional data
}
