% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDDchangepointWBS.R
\name{test_significance_changepoint}
\alias{test_significance_changepoint}
\title{This function conducts the test of significance required for wild binary segmentation}
\usage{
test_significance_changepoint(
  data,
  FUN = dist1,
  cp_estimate = NULL,
  test = "asymptotic",
  npermut = 200,
  siglevel = 0.05
)
}
\arguments{
\item{data}{a matrix of size nxp (i.e., n observations, each p-dimensional)}

\item{FUN}{a dissimilarity measure, default is the new dissimilarity measure d(Xi,Xj)}

\item{cp_estimate}{change point candidate which is calculated if NULL, default is NULL}

\item{test}{either asymptotic or permutation test, default is the asymptotic test}

\item{npermut}{number of permutations for permutation test, default is 200}

\item{siglevel}{the significance level, default is 0.05}
}
\value{
the change point candidate and whether it is significant or not (1 significant, 0 not significant),
           as well as the p-value of the test for change point
}
\description{
This function conducts the test of significance required for wild binary segmentation
}
