% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDDchangepoint.R
\name{test_single_changepoint}
\alias{test_single_changepoint}
\title{This function runs the proposed algorithm for single change point detection in high dimensional data}
\usage{
test_single_changepoint(
  data,
  FUN = dist1,
  test = "asymptotic",
  npermut = 200,
  siglevel = 0.05
)
}
\arguments{
\item{data}{a matrix of size nxp (i.e., n observations, each p-dimensional)}

\item{FUN}{a dissimilarity measure, default is the new dissimilarity measure d(Xi,Xj)}

\item{test}{either asymptotic or permutation test, default is the asymptotic test}

\item{npermut}{number of permutations for permutation test, default is 200}

\item{siglevel}{the significance level, default is 0.05}
}
\value{
the change point candidate and whether it is significant or not (1 significant, 0 not significant),
          as well as the p-value of the test for change point
}
\description{
This function runs the proposed algorithm for single change point detection in high dimensional data
}
