% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDDchangepoint.R
\name{test_statistic}
\alias{test_statistic}
\title{This function calculates the proposed test statistic T for given data and dissimilarity measure}
\usage{
test_statistic(data, FUN = dist1)
}
\arguments{
\item{data}{a matrix of size nxp (i.e., n observations, each p-dimensional)}

\item{FUN}{a dissimilarity measure, default is the new dissimilarity measure d(Xi,Xj)}
}
\value{
the change point estimate together with the test statistics T and W
}
\description{
This function calculates the proposed test statistic T for given data and dissimilarity measure
}
