% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDDchangepointWBS.R
\name{test_statistic_wbs}
\alias{test_statistic_wbs}
\title{This function calculates the proposed test statistic T for given data and dissimilarity measure for wild binary segmentation}
\usage{
test_statistic_wbs(data, FUN = dist1, cp_estimate = NULL)
}
\arguments{
\item{data}{a matrix of size nxp (i.e., n observations, each p-dimensional)}

\item{FUN}{a dissimilarity measure, default is the new dissimilarity measure d(Xi,Xj)}

\item{cp_estimate}{change point candidate which is calculated if NULL, default is NULL}
}
\value{
the change point estimate together with the test statistics T and W
}
\description{
This function calculates the proposed test statistic T for given data and dissimilarity measure for wild binary segmentation
}
